Fix example in README.md
The example from the README file doesn't compile, as `NewSimConnect()`
requires a string argument now. Also the last assignment to `err` is
unchecked, which is a bad idea. And finally the example code can only
work from within this package, but is typically run from the `main`
package.

This PR addresses all these points.

Fixes issue #4.